Updated reset process to not remove the .github folder

+++ build/lib/reset.js
@@ -5,7 +5,7 @@ fs
   .readdirSync(path.resolve('./'))
   .forEach(filepath => {
     const fullpath = path.resolve(filepath)
-    if (filepath.startsWith('.') && filepath !== '.git' && fs.statSync(fullpath).isDirectory()) {
+    if (filepath.startsWith('.') && filepath !== '.github' && filepath !== '.git' && fs.statSync(fullpath).isDirectory()) {
       fs.rmdirSync(filepath, { recursive: true })
       console.log(`Removed ${fullpath}`)
     }
